Add test_selector to tag_helper
As a developer, it can be difficult to write test assertions looking for specific DOM nodes without introducing markup specifically for tests.

Add functionality to display a `data-test-selector` value in non-production environments for the purpose of precisely targeting DOM nodes in test assertions.

We've been using this solution for over four years in the GitHub monolith.

For example:

```erb
<%= tag.p("Hello, world!", test_selector: "greeting") %>
```

```ruby
assert_selector("[data-test-selector='greeting']")
```

I've implemented this functionality as both an option to `tag_helper` and as a standalone `test_selector` view helper. We use both approaches extensively at GitHub.

The behavior is controlled by the `render_test_selectors` configuration flag.

Co-authored-by: Nikka Padilla <nixpad@github.com>