Revert "Add test to make sure this method will not be removed again"
This reverts commit d93a5d385e5bc2392a1f47dc2885e353898b62e1.

Revert "Revert "Remove unused internal methods in ActiveModel::Attributes""

This reverts commit 2d7967204e7f7d5ba846b0a6ed51088c7a7db365.

Reason: read_attribute was added in 6.1 as a performance optimization
and it is not needed anymore and write_attribute only existed to make
possible to call something that is not `attribute=` with send. We don't
need those methods internally and since they were never part of the
public API we can remove them.