Add docs and guards to `while_preventing_writes`
In Rails 6.1 `while_preventing_writes `moved to be called on `Base`
since we no longer have multiple connection_handlers. This change guards
against calling `while_preventing_writes` on the `connection_handler` if
we're not in legacy mode. For the new `while_preventing_writes` we can
call out to the `connection_handler` from there if in legacy mode. We
could do this on the legacy version but I'd like to use this raise as a
kind of hard deprecation. I've also documented the new method to match
docs from the previous method.