Increasing Puma's worker timeout in development
After Puma's version 5, a worker timeout check was added, causing all
debugging sessions (byebug, binding.irb, ...) to be terminated after a
minute is passed.

One minute timeout is fine for production environments but not for
development, when we usually debug stuff and need more than a minute to
do so.

This PR updates Puma's config template file to include a worker timeout
of one hour, only for development environments.