ActionView: FormBuilder default aria-invalid value
When rendering fields for an ActiveModel-enabled form, encode the
validity of the field to the [aria-invalid][] attribute.

Encode the validity for `<input type="text">` descendants (with the
exception of `type="hidden"` and `type="checkbox"`), `<select>`
elements, `<textarea>` elements, and `<trix-editor>` elements.

When rendering an `<input type="radio">`, _only_ render
`[aria-invalid="true"]` for the element marked as `[checked]`.

[aria-invalid]: https://www.w3.org/TR/wai-aria-1.1/#aria-invalid