Link to API docs in AS Core Extensions guide [ci-skip]
This links the first mention of each method to its API documentation,
similar to a Wikipedia article.  Some subsequent mentions are also
linked, when it suits the surrounding text.

This also modifies the text in a few places to ensure that methods are
explicitly mentioned (and linked) before they appear in code examples.