Raise if connection_handlers is called in legacy_connection_handling
While we didn't need to deprecate any behavior when implementing
granular connection swapping it's not 100% clear to the person upgrading
that this needs to change when you switch off
`legacy_connection_handling`. The new version doesn't support
multiple handlers but before it wasn't raising an exception so it was
possible applications were still using it and hitting confusing errors.
This change ensures that if an application is using the new handling
that `connection_handlers` and `connection_handlers=` an exception will
be raised.