Read all records at once when using update with an array
Prior to this change records were loaded one by one and then updated one
by one.

Now records are loaded upfront and updated mostly one by one.  When there are duplicate id's
attributes are merged together ahead of time.  The method
still returns duplicates in the order it was given, meaning:

Model.update [1,1,3], [...]

will still return 3 Model objects in that order.

There are a number of tradeoffs to consider for this PR:

1. Attributes are merged together upfront (so duplicated id's will only
yield one update).  To me this seems better but it is a change
2. We have to rebuild the ordered list with duplicates since
find(ids) removes duplicates
3. We're adding some memory overhead to make O(1) lookups happen