Only use ActionDispatch::ActionableExceptions middleware if local env
When building the default middleware stack, the ActionDispatch::ActionableExceptions
middleware should only be used if the configuration has :consider_all_requests_local
set to true.

There is an existing check #actionable_request? in ActionableExceptions
that will prevent actionable exceptions from being dispatched, which
relies on the "action_dispatch.show_detailed_exceptions" header being
set to false. This check should remain in place as a second line of defense
in case the ActionableExceptions middleware is added back to the application.

However, we can go ahead and remove the middleware from the default stack
when the environment is not local.