Allow custom purpose for ActiveStorage signed ids
This commit allows passing a custom `purpose:` keyword argument to both
`find_signed!` and `signed_id`. This matches the signature of the
`super` methods in ActiveRecord, and it is already possible with the
non-bang [`find_signed`][find_signed] method, since that one doesn't
have an override in ActiveStorage (although perhaps it should, to set
the default :blob_id purpose)

This is useful in cases where you want to further isolate different
types of blobs, for example if they have different levels of
authorization.

[find_signed]: https://github.com/rails/rails/blob/be11d1b6e8853e25fd10f31f1d6422a2c58c06cf/activerecord/lib/active_record/signed_id.rb#L42-L48