Refactor ActiveRecord::Associations::Preloader class
Part 1 of ?

The `ActiveRecord::Associations::Preloader` class was passing around
many of the same arguments to methods and was difficult to follow.

This PR does a few things:

1) Move arguments for the `Preloader` to the initializer instead of
`preload`
2) Merges the `preloaders_for_one` and `preloaders_for_hash` methods
into one `build_preloaders` method that takes no arguments. This new
method is responsible for looping through all the passed associations.
3) If there are nested associations passed, the class will build a new
`Preloader` object for each `child` in `build_child_preloader`.
4) The above changes allowed for removing most of the arguments being
passed down.
5) Implement a deprecated shim that preserves the original API. While
this API is private and undocuented, we know that lots of applications
are using the API. Generally we don't deprecate private APIs but due to
it's wide usage we decided to maintain behavior here.

Next steps:

1) Continue refactoring this class and shared classes as they make more
sense. There's a lot of `flat_map`'s in here that feel overused.
2) Once the API is more reasonable revive older Preloader PRs (ie #32136)
3) Expose a usable public API so that apps can build their own
preloaders when necessary (we do this a lot at GitHub) without
interacting with private methods.

Co-authored-by: Dinah Shi <dinahshi28@gmail.com