Refactor Preloader to remove AlreadyLoaded class
Part 2 of ?

It felt strange to have an `AlreadyLoaded` class when we have 2 classes
for the preloaders. `ThroughAssociation` already super's to
`Association`. Instead of having a new class that provides a confusing
indirection, we can get and set all the information we need on
`Association` inside of `run`.

In addition, the `already_loaded` instance variable on
`ThroughAssociation` actually meant something different than
`already_loaded` that was checked for `AlreadyLoaded`. By pushing this
check down we can get rid of the `initialize` in `ThroughAssociation`
entirely since then all it's doing is calling `super`.

Co-authored-by: John Crepezzi <john.crepezzi@gmail.com>