Add a note about existing apps for ActiveStorage public access.
Every existing file needs to be individually updated to have a
publicly-readable ACL (at least, this is true of S3 and DigitalOcean
Spaces), otherwise the files will fail to load.

This bit me when I was trying to introduce `public: true` in my
production Rails app.

I realize the docs already mention consulting the cloud service docs to
ensure that your bucket is "properly configured for public access", but
that's rather vague and I didn't really understand the potential issues
with individual files' public read permissions from the S3 docs.