Add `primary_key` option to `has_one_attachment` and `has_many_attachments` in ActiveStorage
As it currently stands, there is no way to specify the primary key column
to use when adding ActiveStorage attachments to ActiveRecord models. For
some situations, this is undesirable, as the "primary key" of a specific
model may be different than `id`.

To illustrate, imagine a model `Message` and a model `DataExport`. `Message`
has a primary key `id` of type `uuid`, while `DataExport` has a primary
key `id` of type `bigint`. `Message` also has a `bigint_id` column of type
`bigint`.

If we just put `has_one_attached :file` in both `Message` and `DataExport`,
it obviously wouldn't work because `Message` uses a `uuid` whereas `DataExport`
uses a `bigint`. The solution then is to specify that the primary key column
to use for `Message`'s attachments should be `bigint_id` (assuming it's
unique and not null), but that's not possible without monkey-patching.
This proposal gives the user more fine-grained control while still
defaulting to the convention of `id` as the primary key column.

Example usage is as follows:

``` ruby
class DataExport < ApplicationRecord
  # ...
  has_one_attached :file
  # ...
end
class Message < ApplicationRecord
  # ...
  has_one_attached :file, primary_key: "bigint_id"
  # ...
end
```