Fix rounding NaN and Infinity on newer BigDecimal
Since BigDecimal 2.0.3, BigDecimal#round returns an integer if the
requested precision is less than 1:
https://github.com/ruby/bigdecimal/commit/e8fc98050167fd943574609988b8754414e0a7c1

NaN and Infinity cannot be represented as an integer, so rounding these
values now raises a FloatDomainError. This is a regression of the fix in
081a3963ea2aab617a92874d7ef59ce98eee6a64.

Specifying the rounding mode restores the previous behaviour.

This isn't a problem on master or 6-1-stable, as the rounding mode is
always specified there since 7905bdfd8b2ae50319cd7a9a74ee1f8c865d648d.