Add note for ActionText models that use UUID's
I ran into an issue where I failed to specify `type: :uuid` for the ActionText migration since the migration was automatically generated.

The result was that even after updating the fixtures, my tests were failing due to `nil`  values for the `body` of any rich text attributes. It took a little time to connect the dots, so hopefully this (or something like it) could help save others some time investigating.

Separately, this also made me realize that the polymorphic nature of the `action_text_rich_texts` table means that all models that need rich text attributes would have to use the same approach, either UUID's or integer ID's. This note may be trying to do too much as a result. Very much open to better approaches on how to clarify this to save anyone else some spelunking.