fix ActionView::Base.raise_on_missing_translations
On Ruby 3.0.0, this method is causing the following error:

    ActionView::Template::Error: class variable @@raise_on_missing_translations of ActionView::Base is overtaken by AbstractController::Translation

Since both of them are defining the class variable and Ruby 3.0.0 now makes this into an error. The one in `ActionView::Base` seems to be deprecated now, so I figured it would be safe enough for the `cattr_accessor` to be replaced with a simple delegation to `AbstractController::Translation`.

I have monkey-patched this into my current project and it seems to be working fine.